Ugg and Wrongway
they are both implemented.
It was tricky to find a method for them not to fall of the map unless they were at the end.

GridNavigator was improved to accomodate this.

The probelm now is that the character is not in its exact place in the node. I will find a way to solve this issue momentarely